Add dark mode support and theme toggle with next-themes
- Integrate next-themes for theme management
- Add ThemeProvider and ModeToggle components
- Update layout and pages to support dark mode
- Enhance authentication and landing pages with theme-aware styling
- Refactor middleware and remove unnecessary console logs
