+++ package.json
@@ -32,6 +32,7 @@
     "drizzle-orm": "^0.29.5",
     "lucide-react": "^0.475.0",
     "next": "15.1.6",
+    "next-themes": "^0.4.4",
     "postgres": "^3.4.3",
     "react": "^19.0.0",
     "react-day-picker": "8.10.1",
