+++ src/lib/stack.ts
@@ -2,5 +2,5 @@ import "server-only";
 import { StackServerApp } from "@stackframe/stack";
 
 export const stackServerApp = new StackServerApp({
-    tokenStore: "nextjs-cookie",
+    tokenStore: "nextjs-cookie"
 });
